requirements.txt
streamlit
google-generativeai
google-api-python-client
langchain
langchain-community
langchain-qdrant
qdrant-client
fastembed
google-auth
google-auth-oauthlib
google-auth-httplib2
google-genai
langchain-google-genai
